requirements.txt
fastapi
uvicorn[standard]
scikit-learn
pandas
numpy<2
scikit-surprise
sqlalchemy
pymysql
redis
python-dotenv
pydantic-settings
joblib
httpx
pytest